bluetooth: controller: Moved vendor specific HCI to hal
Moved vendor specific part of hci.c to hal, thereby enabling out-of-tree
HCI implementations.
