+++ subsys/bluetooth/controller/hci/nordic/hal/hci_vendor_hal.h
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2016 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "hal/nrf5/hci.h"
